Started two new packages, one for exploration, another for Webots sim
The Webots_pkg is for all of our webots simulation stuff. This should hopefully make it a ton nicer to work with, as we will control everything within the simulation, and we won't have any weird cross-referencing going on. 

The explore_pkg is VERY unfinished. I'm thinking we can create this package to interface with a separate nav2 package for each individual member of our swarm.